miniscript: implement PartialOrd/Ord nonrecursively
This implements an ordering based on the display ordering, though a
more efficiently than just serializing to strings and comparing those.

Empirically this adds a couple kb to the total code size, but not
as much as the nonrecursive fmt shrunk it. We expect somewhat larger
code to implement such a nontrivial algorithm nonrecursively (and in
fact the old algorithm was "wrong" in that it did not implement any
coherent ordering).

This is a breaking change because it changes the order of Miniscripts.