remove the TranslatePk trait
We actually just empty it and deprecate it in an attempt to minimize
breakage. People *implementing* the trait will break anyway because the
next commit will have me changing the Translator trait. But the typical
case is that somebody imports the trait, then calls .translate_pk on
some object, and we want that usecase to continue working with only
warnings.