checksum: clean up errors and error paths
This creates a new checksum::Error type. For now it sticks it into the
giant global error enum, but we will fold it into an expression-parsing
enum in the following commits.

This eliminates several instances of the Error::BadDescriptor variant,
which is a meaningless stringly-typed variant that we'd like to
eliminate.