introduce `StaticDebugAndDisplay` blanket trait, use it for all errors
We will want to start boxing our errors to avoid generic error types
(and to reduce the size of error types, by collapsing multiple variants
into one). In order to do this, we need a trait that we can box which
expresses fmt::Debug and fmt::Display.

(Ideally we would also bound on std::error::Error but we can't express
this bound without std until Rust 1.81, and we can't have a std-only
bound since it would make the std feature non-additive.)

This IS A BREAKING CHANGE because it now requires all key parsing errors
to have a 'static bound. But the alternatives are pretty bad:

* We could continue dropping parsing errors and replacing them with
  Strings (or not replacing them at all..); or
* We could add a <Pk> bound to all of our error types