remove PartialEq bound from top-level Error type
This might annoy some people, but it will be necessary in order to hold
generic pubkey error types inside of our error returns.

Our alternatives are:

* Implement some manual hacky notion of equality for pubkey error
  (either "never equal" or "equal based on comparing output of
  fmt::Display or something)
* Add a PartialEq bound on all the <Pk as FromStr>::Err types, which
  would make it impossible to use anything that e.g. contained an io
  error.

I don't think ether of these are good, and given that io::Error has "set
the standard" for error types to have basically no useful trait bounds
other than Debug and Display, we are okay to do this.