descriptor: add fuzz test comparing to published rust-miniscript 12
This bumps the local Cargo.toml version to 13, which will be the next
release (since we've made many breaking changes), and in the fuzz test
adds an explicit dependency on miniscript 12 from crates.io, as
`old_miniscript`.

Adds a single fuzztest which attempt to parse descriptors with both
master and 12, to make sure they're the same.