miniscript: make display prefer 'u' over 'l' in the fragment l:0
When fuzzing my new non-recursive tree parsing logic, I noticed that we
were deviating from the released 12.0 in the way that we display l:0.
This is an ambiguous fragment which can be displayed either as l:0 or
u:0. In our released code we use u:0 so stick with that.

This was unintentially changed as part of #722. Change it back.

While we are at it add a regression test for https://github.com/rust-bitcoin/rust-miniscript/pull/735