descriptor: eliminate several instances of Unexpected
When we are parsing keys using from_str, use the appropriate variant of
ParseError.

Also there was one place where we were taking an Error, converting it to
a string, then wrapping in Error::Unexpected. I suspect this was a
rebase mistake or something like that. Just get rid of the conversion.

There are now 2 instances of Error::Unexpected in the codebase and one
instance of errstr.