expression: encapsulate threshold parsing
Our current expression API has a `to_null_threshold` method, which works
reasonably well but requires the caller translate the returned threshold
by accessing individual children.

We will later want to change the Tree represntation to make individual
child access inefficient. To do this, encapsulate the threshold
construction into a new verify_threshold.

This previously was not done because our messy error structures made it
very difficult to manage. But our error structures are less messy so
it's possible now, though a bit of a hack. (We map everything to the
global Error structure. Later we will do pretty-much the same thing
but with ParseError in place of Error, which will be more elegant.)