expression: drop Error::ParseTree variant
The correct way to build a Error::ParseTree is to make an Error::Parse
containing a ParseError::Tree. Eventually when we get rid of the
top-level Error enum this will be more sensible, and we will be able to
use ? instead of this awful `.map_err(From::from).map_err(Error::parse)`
construction.

This commit rightfully belongs as part of the previous commit, but it's
noisy and mechanical so I separated it out.