expression: replace poorly-typed `binary` function with new one
Improves the typing of the errors (and the error messages themselves) by
eliminating another instance of `errstr`.

Saves another instance of `errstr` by removing the unused `unary`
method. There are now 5 left in the entire library :).

Strictly speaking, we ought to do a deprecation cycle on these. But
because we are changing the `expression` module so thoroughly, and we
don't expect any users are directly using it, it doesn't seem worth the
difficult (impossible?) task of preserving the old API for the sake of
deprecation messages.