expression: replace most uses of `terminal`
This is a big diff but it's almost entirely mechanical. Replaces the old
expression::terminal method with two new `verify_terminal_parent` and
`verify_terminal` methods, which do the appropriate checks and return a
strongly-typed error.

Does not directly eliminate any instances of errstr or Unexpected
(though the next commit will), but it does eliminate a large class of
them: now when you try to parse an expression with a bad number of
children, e.g. and_v with 3 children, you will get an error message that
says this rather than an opaque "Unexected(<<and_v>>)" or whatever you
get now.

It does reduce the semantic::PolicyError type to a single variant, which
has no information and is only used to indicate that the entailment
calculation failed. This can be replaced by an Option (and will be, in
the next commit). It also eliminates some uses of concrete::PolicyError,
but the variants are still used by the absurd Policy::is_valid method,
so we have to keep them for now.

Also, you may notice that this commit and others have a ton of calls to
.map_err. I apologize for this. But when we change error types so that
parsing returns a string-parsing-specific error rather than the giant
Error enum, these should mostly go away.