tr: encapsulate object which is yielded by TapTreeIter
All this does is wrap the existing (depth, script) return value in a
structure that has a pile of accessors on it. The accessors include
a couple of expensive computations which are currently done manually.

A later commit/PR will introduce a `SpendInfo` structure which will
cache these computations and amortize much of the hashing needed to
produce control blocks. But these new accessors simplify some things.