compiler: refactor out a call to errstr
The private function `with_huffman_tree` can only fail if it is given an
empty input. In both places we call it, we unwrap the result because we
know this is impossible.

Instead, just change the function to panic internally instead of
returning an error we're just going to unwrap.

Since the error was constructed with errstr, this gets rid of a call to
errstr.