expression: replace methods for parsing numbers and locktimes
This eliminates another instance of `errstr` and provides well-typed
errors when parsing locktimes and threshold values.

It copies the AbsoluteLockTime and RelativeLockTime error variants from
the main Error enum into ParseError. The old copies are now used only in
script decoding, and will be removed in a later PR when we give script
decoding its own error.