clippy: a couple tiny fixes
Clippy prefers we do string slicing after conversion to bytes, not
before, which may be more efficient (at least, in elimantes some panic
paths which we know are impossible since our strings are ASCII, but the
compiler probably doesn't).

It also changes some nested lists in docs to use 1/2/3 numbering because
that's what rustdoc recognizes (and clippy complains about weird
indentation with the existing a/b/c sublist).