fuzz: add miniscript_satisfy target
This commit adds a `miniscript_satisfy` target which includes a fair bit
of infrastructure to generate fuzz data (specifically cheap public keys
which are easy to put into a string, and also a satisfier). Arguably
this stuff should be pulled out of the fuzz test into a library. Punting
on that until later when we move to cargo-fuzz.