ci: update maintainer-tools version
I believe that our previous use of the `rev:` key had no effect. We need
to use `ref:` with a full commit ID. As written, I think that we were
just always pulling master, even though it looked like a commit ID.

Anyway we don't use any cool features of rust-bitcoin-maintainer-tools
so it doesn't really matter here. But we should pin it properly.