tr: optimize parsing of taptrees
This gets our benchmarks to roughly where we were before we flattened
the data structure. The `_tr_oneleaf_` benchmarks appear to be
measurably faster than before but everything else is pretty-much the
same.

Since the old code was definitely slower -- using Arcs, making tons of
tiny allocations, storing internal nodes -- we can infer that actual
construction of TapTrees is just not a meaningful portion of the time
taken to parse expressions, and further optimization effort should go to
the expression parser.