tr: introduce new `TrSpendInfo` structure which holds a full TapTree
This commit introduces a new data structure but **does not** use it. The
next commit will do this. I have separated them so that this one, which
introduces a bunch of algorithmic code, can be reviewed separately from
the API-breaking one.

When computing a `Tr` output, we need to encode all its tapleaves into
Script, put these into a Merkle tree and tweak the internal key with the
root of this tree. When spending from one of the branches of this
output, we need the Merkle path to that output.

We currently do this by using the `TaprootSpendInfo` structure from
rust-bitcoin. This is not a very good fit for rust-miniscript, because
it constructs a map from Tapscripts to their control blocks. This is
slow and memory-wasteful to construct, and while it makes random access
fairly fast, it makes sequential access pretty slow. In Miniscript we
almost always want sequential access, because all of our algorithms are
some form of either "try every possibility and choose the optimum" or
"aggregate every possibility".

It also means that if there are multiple leaves with the same script,
only one copy will ever be accessible. (If they are at different depths,
the low-depth one will be yielded, but if they are at the same depth
it's effectively random which one will get priority.) Having multiple
copies of the same script is a pointless malleability vector, but this
behavior is still surprising and annoying to have to think about.

To replace `bitcoin::TaprootSpendInfo` we create a new structure
`TrSpendInfo`. This structure doesn't maintain any maps: it stores a
full Merkleized taptree in such a way that it can efficiently yield all
of the leaves' control blocks in-order.

It is likely that at some point we will want to upport this, or some
variant of it, into rust-bitcoin, since for typical usecases it's much
faster to use and construct.