tr: move Liftable impl for TapTree
Moving this separately from the rest of the TapTree stuff because I
changed the Policy import to Semantic and the Error import to just
crate::Error (since this is the only place that type is used in the
taptree module, and I don't really like this type).

So this is a move-only diff but it might not look like it, so I figured
I should at least make the diff small.