tr: move TapTree depth-checking to construction rather than embedding in Tr
This makes it impossible to construct a `TapTree` whose depth exceeds
the limit, which means that we can assume that our maximum depth is 128
in all of our code.

Also introduces a new error type for this case distinct from "maximum
recursive depth exceeded" which is supposed to be about limitations of
this library, or at least of Script, rather than about the Taproot
limit.