tr: break Tr API in a couple ways
We fix the return value of Tr::tap_tree to return a Option<&T> instead
of an &Option<T>; we already do this in Descriptor::tap_tree, which is
likely to be more commonly called ... and both methods are pretty
obscure.

Also remove the iter_scripts method in favor of leaves. We did this
rename in the previous PR. However, that PR broke its API anyway by
making it yield an opaque struct rather than a tuple. Better to just do
all the breakage at once.

Maaybe we should backport the rename to 12.x as a pure rename, so that
people will update their code and get nicer error messages.