tr: flatten TapTree type
This replaces the internal representation of `TapTree` with one that
matches the BIP371 iterator: it's a vector of depths and leaves and
nothing else. It does not track its maximum height (except during
construction to return an error if 128 is exceeded); it does not track
"internal nodes"; it does not offer an API to iterate through its
"internal nodes" except by calling Liftable::lift to get a Semantic copy
of the tree.

This greatly simplifies the type and many algorithms associated with it.
It automatically eliminates recursion from display, lifting, dropping,
and so on.

As it turns out, we never use TapTrees except to iterate over the
leaves, and there is no data about internal nodes that we can even
compute, let alone store, so there's no reason to have nodes for them.

In later commit/PR we will create a second TapTree-like type, which can
be constructed from a TapTree for ToPublicKey keys, and can be used to
extract Merkle paths and control blocks. But without ToPublicKey we
can't compute these, so TapTree is not the type for that.