miniscript: add validate() method to Miniscript type
This is a fairly simple commit and maybe should be squashed into the
previous one. It still does not actually use any of the functionality
that is being introduced.

(Trivia: in a much earlier version of this commit, I added a
validation_params field to the Miniscript type, with the idea that
the validation parameters would be set at the terminals (to e.g.
Segwitv0::SANE) and then propagated upward, so that it would be
impossible to construct a Miniscript fragment without having the
validation parameters set.

However, it turned out that this is not really workable, because many of
our parameters (notably allow_sigless_branch and allow_non_b) are "top
level only" checks. If they are set on terminals then they will trigger
incorrectly. So they must be unset, and then they won't propagate to the
top level.

One solution to this could be to separate out the Miniscript type into
two: a recursive "fragment" type used during construction and then a "toplevel"
type which the user actually uses. We have done this for `Expression`
for example. I believe we will eventually get to this point in a later
PR series in which we eliminate recursion from Miniscript and the two
Policy types. But we're not there now, and we can only change one thing
at a time without getting overwhelmed.)