compiler: pull non-binary argument errors from PolicyError into CompilerError
The limitation that we have only binary and()s and or()s is really just
a compiler limit; it's neither needed nor enforced when parsing
policies, printing them, lifting them, etc etc.

Maybe we _want_ to enforce this at the type level, but we don't now, and
given that, these error variants are in the wrong place (and will get in
the way of the refactors in the next commit).