compiler: forbid fragment probabilities of 0
We can cause panics in the compiler comparing OrdF64s which are NaN,
which we can produce by parsing ors where every child has weight 0.

Simply make it a parse error to have any zero-probability branches.
In theory somebody might want an or where some fragments have positive
probability and others are 0, indicating "maximally pessimize these".
But it has never occurred to me to do this, and it's easy to simulate
this by using 100s and 1s, say.