miniscript: use byte arrays in lexer
This is a little less space efficient but it gets rid of a lifetime
parameter on the lexer token, which is annoying because it prevents us
from putting it into error types. (Currently we convert it to a string
before putting it into an error.)

It also eliminates a whole bunch of panics where we convert slices to
hashes and .expect() on the length being right, when we know it's right.